fix(testing,docs): #2090 — apply review feedback (rename .spec.ts, baseline probe, rpID-mismatch race, prose tightening)
Address findings from /pr-review-toolkit on PR #2095:

Test (`e2e/browser/passkey-mfa-gate.spec.ts`):
- Rename `.test.ts` → `.spec.ts` to match the convention every other
  file in `e2e/browser/` follows.
- Probe `/api/v1/admin/audit` BEFORE enrollment and warn loudly when
  the gate was already reachable. Without this, a shared admin with
  a leaked passkey or pre-existing TOTP would silently pass through
  a `passkeyCount > 0` regression — exactly what the spec is meant
  to catch.
- Race the naming-dialog assertion against the destructive-error
  paragraph that `passkey-tile.tsx` renders. An `ATLAS_RPID`
  mismatch (the most plausible local-dev failure) now surfaces as a
  one-sentence diagnostic naming the env var, instead of a 15s
  opaque "naming dialog never visible" timeout.
- Replace `Locator.isVisible({ timeout })` (option is a noop) with
  `Locator.waitFor({ state: "visible", timeout })` in cleanup so a
  slow render doesn't skip cleanup and leak the credential.
- Promote cleanup-failure logs from `console.debug` to
  `console.warn` so they appear in CI's `reporter: "github"` —
  silent cleanup degradation was the file's stated load-bearing
  failure mode.
- Annotate the inline `.catch(() => false)` per CLAUDE.md.
- Suffix the enrolled passkey name with `crypto.randomUUID()` to
  avoid `Date.now()` collisions under `retries: 1`.
- Split the CDP teardown into separate try/catch for
  `removeVirtualAuthenticator`, `WebAuthn.disable`, and
  `cdp.detach()` so each failure is individually attributable.
- Rewrite the docblock to describe durable purpose instead of
  narrating PR/issue history.

Docs (`apps/docs/content/docs/security/passkeys.mdx`):
- Drop hard-coded `:664` / `:673` line-number anchors from
  `server.ts` permalinks. The line numbers will rot the first time
  a plugin is added above them — the surrounding prose names what
  to search for.
- Tighten the Trust Device section: the "Trust this device"
  checkbox only renders on the TOTP sign-in page today; passkey-
  only users have no equivalent opt-in.
- Tighten the `NotAllowedError` prose to distinguish the browser-
  raised error from Better Auth's `result.error` re-surface.
- Replace the closed-issue link in the regional-rpID limitation
  with an inlined rationale.
- Fix the device-name example: `derive-device-name.ts` returns
  `Mac · Safari`, not `MacBook · Safari`.
